Upgrade ProtobufJson to be easier to use.
Previously, ProtobufJson required a verbose incantation starting with
specifying the conversion direction and requiring the exact proto to be
specified.

I personally found it difficult to use, despite being the author.

This commit includes a number of interface changes listed below that are
intended to improve useability. Changes are not backward compatible so
this is a major version change.

 * The build process downloads and builds protobuf, rather than
   requiring the user to manually install on it on their system.
 * `ProtobufJson ToProto` becomes `JsonToProto`.
 * `ProtobufJson ToJson` become `ProtoToJson`.
 * Specifying the exact proto file is no longer required, but for large
   search paths, it can speed up the conversion significantly.
 * Options can be specified by the environment variable
   `PROTO_JSON_OPTIONS`. This is useful to simplify commands if a
   particular search path is used often.